# subscriber: fix incorrect filter selection + improve filter performance