@hawkw 2020-02-14
I believe @samschlegel has also seen _very_ significant reductions in the performance impact of slow filtering in more real-world benchmarks.