@fw42 2024-11-21
I'm seeing this same bug on the latest version. Has there possibly been a regression after this was fixed?